refactor: fix components folder
