+++ src/pages/addCard/components/Dot.tsx
