+++ src/pages/addCard/components/index.ts
@@ -1,7 +1,7 @@
-export { default as AddCardForm } from './addCard/AddCardForm';
-export { default as CardExpiredDateInput } from './addCard/CardExpiredDateInput';
-export { default as CardHolderNameInput } from './addCard/CardHolderNameInput';
-export { default as CardNumberInput } from './addCard/CardNumberInput';
-export { default as CardPasswordInput } from './addCard/CardPasswordInput';
-export { default as CardSecurityCodeInput } from './addCard/CardSecurityCodeInput';
-export { default as Dot } from './addCard/Dot';
+export { default as AddCardForm } from './AddCardForm';
+export { default as CardExpiredDateInput } from './CardExpiredDateInput';
+export { default as CardHolderNameInput } from './CardHolderNameInput';
+export { default as CardNumberInput } from './CardNumberInput';
+export { default as CardPasswordInput } from './CardPasswordInput';
+export { default as CardSecurityCodeInput } from './CardSecurityCodeInput';
+export { default as Dot } from './Dot';
